tests/requirements.txt
pytest==7.0.0
pytest-cov==2.12.1
pytest-logger==0.5.1
segmentation-models-pytorch==0.2.1
